Build block template + Simulator (simpa) (#78)
* Initial generator example

* Initial working framework

* Impl a run function

* Move to sim::infra

* Kaspa network sim

* Running consensus for each miner

* lint

* Enforce max parents rule

* Exclude simpa from main workspace, and make it a separate nightly package

* Move coinbase structs to consensus core

* Block template struct

* Refactor simulator and avoid generators

* Build block template

* Fix initial virtual (over genesis)

* Tx sign + passing test (with Elichai)

* Add ignored big data test

* Add (sk, pk) key pair to each simulation miner

* Measure validation time

* Validate with concurrency

* Mine blocks with transactions

* Refactor + print simulation progress

* Sign txs in parallel

* Command line interface

* Calculate GHOSTDAG K as a function of delay and bps

* Print DAG stats

* Fix an error in mining sampling rate

* Set DAA window size and Merge depth according to BPS

* Minor

* Change defaults to current network params

* use ktype

* Print processing rates explicitly

* Make sure we do not explode memory

* Swap remove

* Perf params with cache sizes from go

* Adapt perf params to simulation requirements

* Bound cache sizes properly

* Use expected_header_pruning_point for build_block_template

* Use past_median_time in block template timestamp calculation

* Bound cache sizes accurately

* Command line option for target number of blocks

* Scale coinbase maturity with bps*delay

* eprint -> print

* Minor

* Use the updated header constructor

* Remove timestamp and nonce from public build template api

* Add a Refs enum to represent a varying pointer type, and use it to simplify traversal_manager refactor

* Use Refs for ghostdag protocol refactor

* Address review comments

+++ math/src/lib.rs
@@ -45,7 +45,7 @@ impl Uint256 {
             compact >>= 8;
             size += 1;
         }
-        compact | (size << 24) as u32
+        compact | (size << 24)
     }
 }
